Test concept from Database to V1 HTTP.
This particular test is checking that /v1/pools is served correctly given
the right data in stake_events database table.

We can write the other v1 functionality tests based on this.

Also adds a visible error message for the case when the LastBlock is called
without setting a block first.