Create test database.
Old database setup was connecting to the production database, and it
was doing every query in a database transaction. That setup was not
able to handle parallel queries in the code. On the long term we plan
the tests not to connect to the production database at all, but to use
the pgtest database.

Change also introduces a new test and corrects an bug introduced in earlier
commit (height was not set properly).