Add a query that checks the recorded rune and asset depths against a derivation of these numbers from the previous depths and events.
This query is useful for documentation purposes because the formula for deriving pool depths can be verified experimentally (by running the query), and can be inspected manually to reason about what is going on.

At time of writing and with the chain synced only to 03.10.2021, all of the blocks in BTC.BTC and BNB.BNB pools add up, i.e. asset_chg_check and rune_chg_check are 0. For ETH.ETH there are 28 blocks that don't add up, and this may be due to missing corrections that are applied in Midgard but not in the formula implemented in this query.