Fix priceShiftLoss when interval is not provided
The beforeMap was not considered if no interval is present.
So the resulting calculation would end up in an NaN. Since
from/to is present, we need to initialize the beforeMap for
both cases, if interval is present and if an interval is not
present. This fixes this issue and adds a testcase for it.